Błąd przy N-numeracji
Gdy nie ma spacji po N30 - nie usuwa n-numeracji

+++ c++ files/gCode Repair/options.cpp
@@ -60,7 +60,7 @@ bool option_1(string line[], unsigned short& line_amout, Console& console)
     middle_y = old_y + old_j;
     //korekcja odchy³u g³owicy #12
     float corr_diametr = 16;
-    float corr_middle_x = (float) ( - 0.5);
+    float corr_middle_x = - 0.5;
     float corr_middle_y = 2;
     
     cout << endl << "Podaj <korekcje srednicy> <korekcje osi x> <korekcje osi y> \nWartosci musza byc oddzielone spacja (np.: 16 -0.5 2)\n Wpisz:";
